Extracted search utils from Aethel lib
